requirements.txt
diffusers
transformers
accelerate
Pillow
numpy
tqdm
clean-fid